requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.6
outcome==1.2.0
packaging==23.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
